v2 Stage A move #2+#3: extract writer + MCP delegates via subprocess
Move #2 — extract writer-side machinery into state/writer.mjs:
  writeState · initState · cleanPipelineScope · archivePreviousRun
  recordFixLoopHistory · sumFixLoopHistoryForCheck · sumFixLoopsForPhase
  deriveLegacyGateBooleans · applyLegacyGateDerivation
  normalizeBlockedHookState · recordRunbookTransition
  UnsupportedSchemaVersionError · VALID_PHASES · DEFAULT_STATE
  PIPELINE_SCOPE_PATHS

  mpl-state.mjs shrinks 863 → ~52 lines — thin facade re-exporting reader
  and writer surface at original names. 33+ caller import sites untouched.

Move #3 — MCP delegates to hooks writer via subprocess shell-out:
  + hooks/lib/state/writer-cli.mjs — JSON-on-stdin → writeState() wrapper
  mcp-server/src/lib/state-manager.ts spawns writer-cli for every state
  mutation. CURRENT_SCHEMA_VERSION on MCP side bumped 2 → 7.
  Non-atomic tmp+rename fallback removed.

  Strategy: subprocess-shellout (not direct-import). Justification: MCP TS
  build avoids cross-module-resolution surface; writer-cli boundary makes
  future flip to dynamic import() a one-file change if subprocess overhead
  (30-80ms per write) ever profiles as a hotspot.

  MPL_HOOKS_ROOT env var supports out-of-tree MCP packaging.

Closes evaluation finding: "MCP writeState skips I5/H8/normalize/caps"
and "schema v2 vs v7 drift" — both eliminated structurally, not patched.

Verify: 1,629 hooks tests pass · 0 regressions · MCP tsc build clean.
